Rename directory to match remaining code
With the v1 metrics code deleted, the only code left is for the
`metrics_transmitter` binary.  Rename the directory to match that
purpose.

Bug: 463443661
