+++ base/cvd/cuttlefish/host/commands/metrics_transmitter/metrics_transmission.h
